rearranged float with double this makes better sense in terms of type ordering

+++ src/Kingdom.Protobuf.Descriptors/Enumerations/ProtoType.cs
@@ -9,12 +9,12 @@ namespace Kingdom.Protobuf
         /// <summary>
         /// 
         /// </summary>
-        Double,
+        Float,
 
         /// <summary>
         /// 
         /// </summary>
-        Float,
+        Double,
 
         /// <summary>
         ///
